+++ widgets/DownloadIcsLink.php
@@ -8,11 +8,7 @@
 namespace humhub\modules\calendar\widgets;
 
 use humhub\components\Widget;
-use humhub\libs\Html;
-use humhub\modules\calendar\helpers\Url;
 use humhub\modules\calendar\interfaces\event\CalendarEventIF;
-use humhub\modules\calendar\models\CalendarEntry;
-use Yii;
 
 /**
  * Class DownloadIcsLink
@@ -28,9 +24,11 @@ class DownloadIcsLink extends Widget
     public function run()
     {
         if ($this->calendarEntry === null) {
-            return;
+            return '';
         }
 
-        return Html::a(Yii::t('CalendarModule.base', 'Download ICS'), Url::toEntryDownloadICS($this->calendarEntry), ['target' => '_blank']);
+        return $this->render('downloadLink', [
+            'calendarEntry' => $this->calendarEntry,
+        ]);
     }
 }